Added a field worker & disjunctive fields
If the SVM is working in 3 dimensions (possibly via multidimensional
scaling), we can compute and visualize local classification fields using
a grid of meshes.  This extension performs the field computations
(updating subfield mesh positions and their predictions) in a separate
thread to keep the UI as freed up as possible.  Thus, there is now a
field thread in addition to the force thread.  Of course, this is on the
path towards calling this thing cljs-force-field.

As alluded to above, there are now separate fields visualized for local
connected components in the network.